PRSD-835: Moves search-related view models into searchResultModels
