+++ src/main/kotlin/uk/gov/communities/prsdb/webapp/models/viewModels/searchResultModels/PropertySearchResultViewModel.kt
@@ -1,4 +1,4 @@
-package uk.gov.communities.prsdb.webapp.models.viewModels
+package uk.gov.communities.prsdb.webapp.models.viewModels.searchResultModels
 
 import uk.gov.communities.prsdb.webapp.database.entity.PropertyOwnership
 import uk.gov.communities.prsdb.webapp.models.dataModels.RegistrationNumberDataModel
